Change the Lauchpad to a new component and update styling and position
